图片更换

+++ src/org/ddpush/client/demo/tcp/Util.java
@@ -20,8 +20,8 @@ public class Util {
 
         StringBuffer sb = new StringBuffer(bytes.length * 2);
         for (int i = 0; i < bytes.length; i++) {
-            sb.append(convertDigit((int) (bytes[i] >> 4)));
-            sb.append(convertDigit((int) (bytes[i] & 0x0f)));
+            sb.append(convertDigit(bytes[i] >> 4));
+            sb.append(convertDigit(bytes[i] & 0x0f));
         }
         return (sb.toString());
 
@@ -31,8 +31,8 @@ public class Util {
 
         StringBuffer sb = new StringBuffer(len * 2);
         for (int i = pos; i < pos+len; i++) {
-            sb.append(convertDigit((int) (bytes[i] >> 4)));
-            sb.append(convertDigit((int) (bytes[i] & 0x0f)));
+            sb.append(convertDigit(bytes[i] >> 4));
+            sb.append(convertDigit(bytes[i] & 0x0f));
         }
         return (sb.toString());
